requirements.txt
rel>=0.4.9.3
langchain>=0.0.0
websocket-client>=1.7.0
openai-whisper>=20231117
async-timeout>=4.0.3
asyncclick>=8.1.3.4
kasa-crypt>=0.4.1
orjson>=3.9.10
python-kasa>=0.5.4
